New grouping after merge
