--- app/src/main/java/com/nervousfish/nervousfish/database/IDatabase.java
@@ -1,9 +1,0 @@
-package com.nervousfish.nervousfish.database;
-
-import java.security.KeyPair;
-
-public interface IDatabase {
-
-    boolean hasKeyPair();
-    KeyPair readKeyPair();
-}
