+++ app/src/main/java/com/nervousfish/nervousfish/modules/cryptography/KeyGeneratorAdapter.java
@@ -4,6 +4,14 @@ import com.nervousfish.nervousfish.service_locator.IServiceLocator;
 import com.nervousfish.nervousfish.service_locator.IServiceLocatorCreator;
 import com.nervousfish.nervousfish.service_locator.ModuleWrapper;
 
+import java.security.KeyFactory;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.NoSuchAlgorithmException;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.RSAPrivateKeySpec;
+import java.security.spec.RSAPublicKeySpec;
+
 /**
  * An adapter to the default Java class for generating keys
  */
@@ -27,5 +35,55 @@ public final class KeyGeneratorAdapter implements IKeyGenerator {
      */
     public static ModuleWrapper<KeyGeneratorAdapter> newInstance(final IServiceLocatorCreator serviceLocatorCreator) {
         return new ModuleWrapper<>(new KeyGeneratorAdapter(serviceLocatorCreator));
+    }
+
+    /**
+     * Generates a random KeyPair with the RSA algorithm.
+     * @return a randomly generated KeyPair
+     */
+    public static KeyPair generateRandomKeyPair() {
+        try {
+            KeyPairGenerator k = KeyPairGenerator.getInstance("RSA");
+            k.initialize(2048);
+
+            return k.generateKeyPair();
+        } catch (NoSuchAlgorithmException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    /**
+     * Gets from the parameter 'kp' the public key and makes a RSAPublicKeySpec out of it.
+     * @param kp the KeyPair with the public key
+     * @return the RSAPublicKeySpec
+     */
+    public static RSAPublicKeySpec getPublicKeySpec(KeyPair kp) {
+        try {
+            KeyFactory fact = KeyFactory.getInstance("RSA");
+
+            return fact.getKeySpec(kp.getPublic(),
+                    RSAPublicKeySpec.class);
+        } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    /**
+     * Gets from the parameter 'kp' the private key and makes a RSAPrivateKeySpec out of it.
+     * @param kp the KeyPair with the private key
+     * @return the RSAPrivateKeySpec
+     */
+    public static RSAPrivateKeySpec getPrivateKeySpec(KeyPair kp) {
+        try {
+            KeyFactory fact = KeyFactory.getInstance("RSA");
+
+            return fact.getKeySpec(kp.getPrivate(),
+                    RSAPrivateKeySpec.class);
+        } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
+            e.printStackTrace();
+        }
+        return null;
     }
 }
